Refactor: remove `LeaderState`.
- Refactor: remove `LeaderState`:
  A single `leader_loop()` is quite enough.  Move leader specific
  methods to `RaftCore`, so that `RaftCore` as a `Runtime` do not have
  to distiguish server-states.

- Refactor: remove replication channel:
  `ReplicationStream` sends message to `RaftCore` via `tx_api`.  Do not
  use a dedicated channel any more. This is one of the steps to finally
  letting `RaftCore` have only one input channel, to avoid creating a
  future for `tokio::select!`.
  `ReplicaEvent` is merged into `RaftMsg`.
