+++ openraft/src/core/mod.rs
@@ -1,11 +1,12 @@
-//! The core logic of a Raft node.
+//! The `RaftCore` is a `Runtime` supporting the raft algorithm implementation [`Engine`].
+//!
+//! It passes events from an application or timer or network to `Engine` to drive it to run.
+//! Also it receives and execute [`Command`] emitted by `Engine` to apply raft state to underlying storage or forward
+//! messages to other raft nodes.
 
-mod admin;
 mod append_entries;
-mod client;
 mod install_snapshot;
 mod internal_msg;
-mod leader_state;
 mod raft_core;
 pub(crate) mod replication;
 mod replication_expectation;
@@ -14,7 +15,6 @@ mod server_state;
 mod snapshot_state;
 
 pub(crate) use internal_msg::InternalMessage;
-use leader_state::LeaderState;
 pub use raft_core::RaftCore;
 pub(crate) use replication_expectation::Expectation;
 pub(crate) use replication_state::replication_lag;
