+++ openraft/tests/elect/t10_elect_compare_last_log.rs
@@ -40,10 +40,14 @@ async fn elect_compare_last_log() -> Result<()> {
         })
         .await?;
 
-        sto0.append_to_log(&[&blank(0, 0), &Entry {
-            log_id: LogId::new(LeaderId::new(2, 0), 1),
-            payload: EntryPayload::Membership(Membership::new(vec![btreeset! {0,1}], None)),
-        }])
+        sto0.append_to_log(&[
+            //
+            &blank(0, 0),
+            &Entry {
+                log_id: LogId::new(LeaderId::new(2, 0), 1),
+                payload: EntryPayload::Membership(Membership::new(vec![btreeset! {0,1}], None)),
+            },
+        ])
         .await?;
     }
